+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `Builder.configDirectory` option to let `Generator` output files that GraalVM needs for AOT compilation ([issue eclipse/deeplearning4j#7362](https://github.com/eclipse/deeplearning4j/issues/7362))
  * Fix `Parser` error on `template<>` containing non-type parameters without names ([issue bytedeco/javacpp-presets#889](https://github.com/bytedeco/javacpp-presets/issues/889))
  * Bundle also the `vcruntime140_1.dll` and `msvcp140_1.dll` redist files from Visual Studio
  * Fix `Builder` for different "java.home" path returned by latest JDKs from Oracle ([pull #400](https://github.com/bytedeco/javacpp/pull/400))